Move FolderMapper to OCA\Mail\IMAP namespace
