+++ tests/Integration/Service/FolderMapperIntegrationTest.php
@@ -24,7 +24,7 @@ namespace OCA\Mail\Tests\Integration\Service;
 use Horde_Imap_Client_Socket;
 use OCA\Mail\Account;
 use OCA\Mail\Folder;
-use OCA\Mail\Service\FolderMapper;
+use OCA\Mail\IMAP\FolderMapper;
 use PHPUnit_Framework_TestCase;
 
 class FolderMapperIntegrationTest extends PHPUnit_Framework_TestCase {
